fix: исправить работу с жанрами и рейтингом

+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -24,7 +24,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre getGenreById(@PathVariable Long id) {
+    public Genre getGenreById(@PathVariable Integer id) {
         return genreService.getGenreById(id);
     }
 }
